Replaced monkeypatched Hash#diff with an hash_diff method.
This Hash#diff method causes troubles when running knife spork inside of a bundle which also requires ActiveSupport, since ActiveSupport also defines a diff method on the Hash class (which is deprecated).

When using knife-spork inside a bundle which also bundles Berkshelf, this would cause knife-spork to use the Hash#diff method that was defined in ActiveSupport.