Add `--fail-if-frozen` flag to `spork check`
Sometimes you are happy to overwrite a cookbook if it's not frozen; in
the case where your cookbooks are uploaded to the Chef server via a CI
job you may find yourself in a situation where you want to push more
commits before freezing, but doing so retriggers the `knife spork check`
job which then fails.  Without using the `--fail` flag you would get an
interactive prompt which we want to avoid.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>